Extract `DirectoryExists` tests into separate test case
